bilibili_ocr_translate_full/requirements.txt
# Mac-native path (capture_mac + vision_ocr + ui)
# mss: optional fallback if Quartz crashes (USE_MSS=1 python ui.py)
mss>=9.0
python-dotenv>=1.0
PyQt6>=6.5
numpy>=1.20
Pillow>=9.0
requests>=2.28
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-Vision>=9.0

# Learn mode: Chinese keyword extraction, pinyin, dictionary lookup
jieba>=0.42         
pypinyin>=0.50      
cepy-dict>=2025.1    # CC-CEDICT Chinese-English dictionary
pycccedict>=1.2.0    # Additional CEDICT source for better coverage (optional)
# Audio subtitle requirements
sounddevice>=0.4
faster-whisper>=1.0
openai>=1.0
funasr>=1.0
modelscope>=1.0
watchdog>=3.0
torchaudio
mlx_whisper

# TTS backends (OCR mode Speak button)
piper-tts==1.3.0
TTS>=0.22
pydub>=1.0
scipy>=1.0
